+++ packages/web/src/components/providers/modal-provider.tsx
@@ -1,28 +1,42 @@
 'use client'
 
-import { createContext, useState } from 'react'
+import { createContext, ReactNode, useState } from 'react'
+import ModalClient from '../modal/modalClient'
 
 interface ModalContextValue {
   isModalOpen: boolean
-  openModal: () => void
+  modalContent: ReactNode
+  openModal: (content: ReactNode) => void
   closeModal: () => void
 }
 
 export const ModalContext = createContext<ModalContextValue>({
   isModalOpen: false,
+  modalContent: null,
   openModal: () => {},
   closeModal: () => {},
 })
 
 export const ModalProvider = ({ children }: { children: React.ReactNode }) => {
   const [isModalOpen, setIsModalOpen] = useState(false)
+  const [modalContent, setModalContent] = useState<ReactNode | null>(null)
 
-  const openModal = () => setIsModalOpen(true)
-  const closeModal = () => setIsModalOpen(false)
+  const openModal = (content: ReactNode) => {
+    setModalContent(content)
+    setIsModalOpen(true)
+  }
+
+  const closeModal = () => {
+    setIsModalOpen(false)
+    setModalContent(null)
+  }
 
   return (
-    <ModalContext.Provider value={{ isModalOpen, openModal, closeModal }}>
+    <ModalContext.Provider
+      value={{ isModalOpen, modalContent, openModal, closeModal }}
+    >
       {children}
+      {isModalOpen && <ModalClient>{modalContent}</ModalClient>}
     </ModalContext.Provider>
   )
 }
