+++ packages/web/src/containers/additional-details/AdditionalDetailsModal.tsx
@@ -1,0 +1,11 @@
+'use client'
+
+import { useContext } from 'react'
+
+import { ModalContext } from '@/components/providers/modal-provider'
+
+export default function AdditionalDetailsModal() {
+  const { isModalOpen, closeModal } = useContext(ModalContext)
+
+  return null
+}
